Provide a pyproject.toml file
This should fix various installation problems relaated to requests by properly providing the build requirements.
We are now using setup.cfg, setup.py, requirements.in, requirements.txt and pyproject.toml simultaneously, which is a bit of a mess.
However, this was the least amount of changes needed to provide pyproject.toml .